+++ app/unit-tests/src/org/commcare/android/tests/formentry/IdentityCalloutTests.kt
@@ -16,8 +16,8 @@ import org.commcare.android.resource.installers.XFormAndroidInstaller
 import org.commcare.android.util.ActivityLaunchUtils
 import org.commcare.android.util.TestAppInstaller
 import org.commcare.android.util.TestUtils
-import org.commcare.commcaresupportlibrary.BiometricIdentifier
-import org.commcare.commcaresupportlibrary.BiometricUtils
+import org.commcare.commcaresupportlibrary.identity.BiometricIdentifier
+import org.commcare.commcaresupportlibrary.identity.BiometricUtils
 import org.commcare.commcaresupportlibrary.identity.IdentityResponseBuilder
 import org.commcare.commcaresupportlibrary.identity.model.IdentificationMatch
 import org.commcare.commcaresupportlibrary.identity.model.MatchResult
@@ -41,6 +41,8 @@ class IdentityCalloutTests {
     @JvmField
     var intentsRule = ActivityScenarioRule(FormEntryActivity::class.java)
 
+    var templates : HashMap<BiometricIdentifier, ByteArray> = HashMap(2)
+
     @Before
     fun setup() {
         Intents.init()
@@ -48,6 +50,12 @@ class IdentityCalloutTests {
         TestAppInstaller.installAppAndLogin(
                 "jr://resource/commcare-apps/identity_callouts/profile.ccpr",
                 "test", "123")
+
+        // Initialize biometric templates
+        templates[BiometricIdentifier.LEFT_INDEX_FINGER] =
+            byteArrayOf(0, 0, -21, -67, 0, -64, 25, 62, -69, -124, -91, 29, -50, -107, 58)
+        templates[BiometricIdentifier.LEFT_MIDDLE_FINGER] =
+            byteArrayOf(122, -91, 114, 62, 107, -95, -69, 28, 110, 123, 72, 71, -86, -117, 126)
     }
 
     @After
@@ -102,13 +110,6 @@ class IdentityCalloutTests {
     fun testRegistrationWithTemplates() {
         val formEntryActivity = ActivityLaunchUtils.launchFormEntry("m0-f1")
 
-        var templates : HashMap<BiometricIdentifier, ByteArray> =
-            HashMap<BiometricIdentifier, ByteArray>(2)
-        templates.put(BiometricIdentifier.LEFT_INDEX_FINGER,
-            byteArrayOf(0, 0, -21, -67, 0, -64, 25, 62, -69, -124, -91, 29, -50, -107, 58))
-        templates.put(BiometricIdentifier.LEFT_MIDDLE_FINGER,
-            byteArrayOf(122, -91, 114, 62, 107, -95, -69, 28, 110, 123, 72, 71, -86, -117, 126))
-
         intendRegistrationWithTemplatesIntent()
         performIntentCallout(formEntryActivity)
         TestUtils.assertFormValue("/data/guid", "test-case-unique-guid")
@@ -141,13 +142,6 @@ class IdentityCalloutTests {
     }
 
     private fun intendRegistrationWithTemplatesIntent() {
-        var templates : HashMap<BiometricIdentifier, ByteArray> =
-            HashMap<BiometricIdentifier, ByteArray>(2)
-        templates.put(BiometricIdentifier.LEFT_INDEX_FINGER,
-            byteArrayOf(0, 0, -21, -67, 0, -64, 25, 62, -69, -124, -91, 29, -50, -107, 58))
-        templates.put(BiometricIdentifier.LEFT_MIDDLE_FINGER,
-            byteArrayOf(122, -91, 114, 62, 107, -95, -69, 28, 110, 123, 72, 71, -86, -117, 126))
-
         val registration = IdentityResponseBuilder
             .registrationResponse("test-case-unique-guid", templates)
             .build()
